pyproject.toml
[project]
name = "videoindex"
version = "0.1.0"
description = "Modular video indexing and search system using FastVLM vs Quen2 with vector database storage"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "moviepy>=1.0.3",
    "scenedetect>=0.6.2",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "typer>=0.9.0",
    "pathlib>=1.0.1",
    "dataclasses-json>=0.6.0",
    "timm>=1.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
videoindex = "videoindex.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
